requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
aiohttp==3.9.1
psycopg2-binary==2.9.9 